refactor: remove attributes package
