+++ packages/api-headless-cms/tsconfig.build.json
@@ -10,7 +10,6 @@
     { "path": "../api-tenancy/tsconfig.build.json" },
     { "path": "../api-upgrade/tsconfig.build.json" },
     { "path": "../error/tsconfig.build.json" },
-    { "path": "../attributes/tsconfig.build.json" },
     { "path": "../models/tsconfig.build.json" },
     { "path": "../handler/tsconfig.build.json" },
     { "path": "../handler-aws/tsconfig.build.json" },
