+++ packages/api-headless-cms/tsconfig.json
@@ -10,7 +10,6 @@
     { "path": "../api-tenancy" },
     { "path": "../api-upgrade" },
     { "path": "../error" },
-    { "path": "../attributes" },
     { "path": "../models" },
     { "path": "../handler" },
     { "path": "../handler-aws" },
@@ -46,8 +45,6 @@
       "@webiny/api-upgrade": ["../api-upgrade/src"],
       "@webiny/error/*": ["../error/src/*"],
       "@webiny/error": ["../error/src"],
-      "@webiny/attributes/*": ["../attributes/src/*"],
-      "@webiny/attributes": ["../attributes/src"],
       "@webiny/models/*": ["../models/src/*"],
       "@webiny/models": ["../models/src"],
       "@webiny/handler/*": ["../handler/src/*"],
