--- packages/attributes/.babelrc.js
@@ -1,1 +1,0 @@
-module.exports = require("../../.babel.node")({ path: __dirname });
