--- packages/attributes/CHANGELOG.md
@@ -1,1 +1,0 @@
-# Change Log
