--- packages/attributes/README.md
@@ -1,20 +1,0 @@
-# @webiny/attributes
-
-[![](https://img.shields.io/npm/dw/@webiny/attributes.svg)](https://www.npmjs.com/package/@webiny/attributes)
-[![](https://img.shields.io/npm/v/@webiny/attributes.svg)](https://www.npmjs.com/package/@webiny/attributes)
-[![code style: prettier](https://img.shields.io/badge/code_style-prettier-ff69b4.svg?style=flat-square)](https://github.com/prettier/prettier)
-[![PRs Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg?style=flat-square)](http://makeapullrequest.com)
-
-## Install
-
-```
-yarn add @webiny/attributes
-```
-
-
-## Testing
-
-### Command
-````
-yarn test packages/attributes
-````
