--- packages/attributes/jest.config.js
@@ -1,5 +1,0 @@
-const base = require("../../jest.config.base");
-
-module.exports = {
-    ...base({ path: __dirname })
-};
