--- packages/attributes/src/index.ts
@@ -1,9 +1,0 @@
-export * from "~/Attributes";
-export * from "~/Attribute";
-export * from "~/BooleanAttribute";
-export * from "~/DateTimeAttribute";
-export * from "~/ListAttribute";
-export * from "~/NumberAttribute";
-export * from "~/ObjectAttribute";
-export * from "~/StringAttribute";
-export * from "~/factories";
