--- packages/attributes/tsconfig.build.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "include": ["src"],
-  "references": [{ "path": "../error/tsconfig.build.json" }],
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist",
-    "declarationDir": "./dist",
-    "paths": { "~/*": ["./src/*"] },
-    "baseUrl": "."
-  }
-}
