--- packages/attributes/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["src", "__tests__/**/*.ts"],
-  "references": [{ "path": "../error" }],
-  "compilerOptions": {
-    "rootDirs": ["./src", "./__tests__"],
-    "outDir": "./dist",
-    "declarationDir": "./dist",
-    "paths": {
-      "~/*": ["./src/*"],
-      "@webiny/error/*": ["../error/src/*"],
-      "@webiny/error": ["../error/src"]
-    },
-    "baseUrl": "."
-  }
-}
