+++ packages/models/package.json
@@ -14,7 +14,7 @@
     "directory": "dist"
   },
   "dependencies": {
-    "@webiny/attributes": "^5.21.0"
+    "@webiny/error": "^5.21.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.12.10",
