+++ packages/models/src/attributes/BooleanAttribute.ts
@@ -1,4 +1,4 @@
-import { Attribute, AttributeParams } from "~/Attribute";
+import { Attribute, AttributeParams } from "./Attribute";
 
 export type BooleanAttributeParams = Omit<AttributeParams, "type">;
 export class BooleanAttribute extends Attribute {
