+++ packages/models/src/attributes/DateTimeAttribute.ts
@@ -1,4 +1,4 @@
-import { Attribute, AttributeParams } from "~/Attribute";
+import { Attribute, AttributeParams } from "./Attribute";
 
 export type DateTimeAttributeParams = Omit<AttributeParams, "type">;
 export class DateTimeAttribute extends Attribute {
