+++ packages/models/src/attributes/ObjectAttribute.ts
@@ -1,4 +1,4 @@
-import { Attribute, AttributeParams } from "~/Attribute";
+import { Attribute, AttributeParams } from "./Attribute";
 
 export type ObjectAttributeParams = Omit<AttributeParams, "type">;
 export class ObjectAttribute extends Attribute {
