+++ packages/models/src/attributes/factories.ts
@@ -1,10 +1,10 @@
-import { AttributeParams } from "~/Attribute";
-import { StringAttribute } from "~/StringAttribute";
-import { BooleanAttribute } from "~/BooleanAttribute";
-import { DateTimeAttribute } from "~/DateTimeAttribute";
-import { ListAttribute } from "~/ListAttribute";
-import { NumberAttribute } from "~/NumberAttribute";
-import { ObjectAttribute } from "~/ObjectAttribute";
+import { AttributeParams } from "./Attribute";
+import { StringAttribute } from "./StringAttribute";
+import { BooleanAttribute } from "./BooleanAttribute";
+import { DateTimeAttribute } from "./DateTimeAttribute";
+import { ListAttribute } from "./ListAttribute";
+import { NumberAttribute } from "./NumberAttribute";
+import { ObjectAttribute } from "./ObjectAttribute";
 
 export type CreateAttributeParams = Omit<AttributeParams, "type">;
