+++ packages/models/src/model/factory.ts
@@ -1,0 +1,5 @@
+import { Model, ModelParams } from "./Model";
+
+export const createModel = (params: ModelParams): Model => {
+    return new Model(params);
+};
